Refactor collision detection algorithm to use double precision for velocity calculations

+++ include/collisionDetectionAlgorithm.h
@@ -11,7 +11,7 @@
 // Function prototypes
 void vibrateMotorsAsTask(void *pvParameters);
 void playMP3AsTask(void *pvParameters);
-void calculateStepCountAndSpeed(const SensorData& sensorData, int* stepCount, float* velocity, float userHeight);
-double collisionDetector(const SensorData& sensor_data, const systemSettings& system_settings, float* velocity);
+void calculateStepCountAndSpeed(const SensorData& sensorData, int* stepCount, double* velocity, float userHeight);
+double collisionDetector(const SensorData& sensor_data, const systemSettings& system_settings, double* velocity);
 void collisionAlert(const systemSettings& system_settings, const MP3& mp3, vibrationMotor& vibration_motor, String vibration_pattern);
 #endif
